changements divers
